Refactor: path alias 추가 및 타입 분리
- user.mock.ts 에 있던 타입을 user.type.ts로 이동
- path alias 추가
    - @@types
    - @apis

@types로 할 경우 node_module의 @types와 충돌하여 @@types로 함
