+++ src/hooks/useUser.ts
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
-import { fetchUserInfo } from "../apis/user.api";
-import { UserInfoResponse } from "../mock/user.mock";
+import { fetchUserInfo } from "@apis/user.api";
+import { UserInfoResponse } from "@@types/user.type";
 
 export const useUser = () => {
     const [user, setUser] = useState<UserInfoResponse>({
