+++ src/svg.d.ts
