+++ src/types/mainPage.type.ts
