+++ src/types/user.type.ts
@@ -1,0 +1,14 @@
+export interface LoginResponse {
+    token: string;
+}
+
+export interface UserInfoResponse {
+    name: string;
+    profileImageURL: string;
+}
+
+export interface SubscriptionResponse {
+    channelName: string;
+    channelEmail: string;
+    profileImageURL: string;
+}
