+++ src/types/videoCard.type.ts
@@ -1,7 +1,7 @@
-import { Video } from "./mainPage.types";
+import { Video } from "./mainPage.type";
 
 export interface VideoCardProps {
     video: Video;
     onClick?: (event: React.MouseEvent<HTMLDivElement>) => void;
-    size?: 'small' | 'medium' | 'large';
+    size?: "small" | "medium" | "large";
 }
